LIBHYDRA-465. De-duplicated STOMP message creation for tests
Moved duplicated methods for creating STOMP messages for use in tests
from individual tests to a "test_stomp_message_helper" class, that
can be "require"d by the test files that use them.

https://issues.umd.edu/browse/LIBHYDRA-465
